backend/resume_service/requirements.txt
fastapi==0.109.1
uvicorn==0.29.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.2
aiofiles==23.2.1
pypdf==3.17.1
mammoth==1.8.0
python-docx==1.0.1
tiktoken==0.6.0
pandas==2.2.0
openai==1.14.1
httpx==0.26.0